backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
idna==3.10
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
spotipy==2.24.0
sqlparse==0.5.1
urllib3==2.2.3
